ARM-2 Leg-2: relocate 845-LoC example harness → #[ignore]d integration tests
The native-MXFP4 (E8M0) numeric gate lived as examples/arm2_leg2.rs (845 LoC),
which trips the ≤500-LoC .rs cap (file-size-cap.yml). Relocate it as the proper
regression net for the E8M0 format path, split under the cap:

  tests/arm2_common/support.rs   (391) — shared PRNG/quant/upload/compare/launch
  tests/arm2_leg2_decode.rs      (269) — Family A: CHECK 1 host-GEMV, 2 RIDER-A3, 3 RIDER-A4
  tests/arm2_leg2_prefill.rs     (256) — Family B: CHECK 4 bit-exact, 5 multi-expert memcheck

Each CHECK is a #[test] #[ignore] fn (Requires GB10 GPU). CI builds+links them
against the libcuda stubs (catches kernel-signature drift) but skips execution;
run on a GB10 host with `cargo test -p spark-model --test arm2_leg2_* -- --ignored`.
Check bodies transcribed verbatim (same kernels, arithmetic, launches).
